rename realityDto to realityResponse
