+++ src/main/java/com/example/demo/ports/RetrieveRealitiesOutputPort.java
@@ -1,6 +1,6 @@
 package com.example.demo.ports;
 
-import com.example.demo.entities.RealityDTO;
+import com.example.demo.inputs.models.RealityResponse;
 import com.example.demo.reality.RealityNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,7 +9,7 @@ import org.springframework.http.ResponseEntity;
 import java.util.List;
 
 public interface RetrieveRealitiesOutputPort {
-    List<RealityDTO> getRealities();
-    Page<RealityDTO> getPage(Pageable page) throws RealityNotFoundException;
+    List<RealityResponse> getRealities();
+    Page<RealityResponse> getPage(Pageable page) throws RealityNotFoundException;
     <T> ResponseEntity<T> getRealityById(Long realityId) throws RealityNotFoundException;
 }
